refactor(cmd): move ReplayRunner to its own file
Address review: keep tracee.go focused on TraceeRunner by moving
ReplayRunner, its Run method and createReplayPrinter to pkg/cmd/replay.go.
